requirements.txt
mcp>=0.1.0
requests>=2.25.0
python-dotenv>=0.19.0